pyproject.toml
[tool.poetry]
name = "cotacoes-amigoz-luiz-felipe"
version = "0.1.0"
description = ""
authors = ["Felipe Cunha <lfcunha.ti@gmail.com>"]
readme = "README.md"
packages = [
    { include = "cotacoes" }
]

[tool.poetry.dependencies]
python = ">=3.13"
django = ">=5.2.3,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
drf-spectacular = ">=0.28.0,<0.29.0"
httpx = ">=0.28.1,<0.29.0"
python-decouple = ">=3.8,<4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"



